Use File.join to ensure a single trailing slash
Related project updates perform a find with an id with a trailing slash,
such as 'project_updates/233/'.  Previously, we'd append another
trailing slash and it would be up to the underlying faraday code to fix it.

On ruby 2.5.x, faraday would receive two trailing slashes and not fix
it.  On ruby 2.4.x, we'd send two trailing slashes to faraday
but it would perform the get with only a single trailing slash.

It's unclear what changed in ruby 2.4 vs. 2.5 to behave differently with the same
faraday and other gems, but it's better to ensure we don't try to send an invalid
URI to faraday.